LS-2165 - upgrade to springboot 3.0  * including flyway changes (baseline migration from maven)   * tsid rather than sequences   * new encryption - will need to kick everybody out and clear device table upon update.

+++ listshop-lmt/listshop-data/src/main/java/com/meg/listshop/lmt/data/repository/impl/CustomListLayoutRepositoryImpl.java
@@ -81,9 +81,9 @@ public class CustomListLayoutRepositoryImpl implements CustomListLayoutRepositor
         Query query = entityManager.createNamedQuery("ListLayoutCategoryEntity.defaultCategoryForSiblings");
         query.setParameter("sibling_tags", siblingIds);
         query.setMaxResults(1);
-        List<BigInteger> resultList = query.getResultList();
+        List<Long> resultList = query.getResultList();
         if (!resultList.isEmpty() ) {
-            return resultList.get(0).longValue();
+            return resultList.get(0);
         }
         return null;
     }
